pyproject.toml
[project]
name = "concierge"
version = "0.1.0"
description = "A tool to provision and manage charm development environments"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "Apache-2.0"}
authors = [
    {name = "Canonical Ltd."}
]
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ty>=0.0.1a34",
    "ruff>=0.2.0",
]

[project.scripts]
concierge = "concierge.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/concierge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "DTZ", "T10", "ISC", "ICN", "PIE", "PT", "RET", "SIM", "ARG", "PTH", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = [
    "ANN401",    # Allow Any type annotations
    "PLR0913",   # Allow many function arguments (CLI commands)
    "PLR0912",   # Allow many branches in functions
    "PLR2004",   # Allow magic values in comparisons
    "TRY003",    # Allow long exception messages
    "TRY004",    # Allow ValueError instead of TypeError for type checks
    "TRY300",    # Allow return in try block
    "TRY301",    # Allow raising in conditionals
    "TRY400",    # Allow logging.error instead of logging.exception
    "TRY002",    # Allow raising Exception instead of custom exceptions
]

[tool.ty]
